Moved existing tests to correct folders
