--- __tests__/.eslintrc
@@ -1,13 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "env": {
-    "es6": true,
-    "jest": true,
-    "browser": true
-  },
-  "extends": "../.eslintrc",
-  "rules": {
-    "no-undefined": 0,
-    "react/prop-types": 0
-  }
-}
